requirements.txt
# --- Web Scraping ---
playwright==1.49.0
playwright-stealth==2.0.2

# --- Data Engineering ---
pandas==2.2.3
pycountry-convert==0.7.2
python-dotenv==1.0.1

# --- Database ---
# sqlite3 is built into Python stdlib

# --- Embeddings & Vector Store ---
sentence-transformers==3.3.1
chromadb==0.6.3

# --- LLM Clients ---
openai==1.61.0

# --- EDA & Visualization ---
plotly==5.24.1
seaborn==0.13.2
matplotlib==3.9.4
jupyter==1.1.1
ipykernel==6.29.5
nbformat==5.10.4

# --- Streamlit UI ---
streamlit==1.42.0

# --- Utilities ---
numpy>=1.26.4
scikit-learn>=1.5.2
tqdm==4.67.1
